modified directory paths for homepage modified json format
