+++ src/app/home/area-person-summary/area-person-summary.component.spec.ts
