+++ src/app/home/area-skills/area-skills.component.ts
