+++ src/app/home/area-work/area-work.component.css
